Add auth check to update and delete

+++ controllers/feed.js
@@ -126,6 +126,11 @@ exports.updatePost = (req, res, next) => {
           error.statusCode = 404;
           throw error;
         }
+        if(post.creator.toString() !== req.userId){
+          const error = new Error('Not authorized!');
+          error.statusCode = 403;
+          throw error;
+        }
         if (imageUrl !== post.imageUrl) {
           clearImage(post.imageUrl);
         }
@@ -154,6 +159,12 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
               }
+            console.log(post.creator.toString() + ' - ' + req.userId);
+            if(post.creator.toString() !== req.userId){
+              const error = new Error('Not authorized!');
+              error.statusCode = 403;
+              throw error;
+            }
             //check logged in user
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
